Sample_Py/requirements.txt
Django>=5.2,<5.3
psycopg2-binary>=2.9.6
gunicorn>=21.2.0
python-dotenv>=1.0.0
django-ninja>=1.0.1
pydantic>=2.0.0
requests>=2.28.2
python-dotenv>=1.0.0
PyYAML>=6.0
